Create reuseable inventory components
