+++ src/components/ui/InventoryTabContent.tsx
@@ -17,6 +17,7 @@ import { Inventory, TabItems } from "./InventoryItems";
 import { hasBoost } from "features/game/lib/boosts";
 import { getCropTime } from "features/game/events/plant";
 import { getKeys } from "features/game/types/craftables";
+import { getShortcuts } from "features/farming/hud/lib/shortcuts";
 
 const ITEM_CARD_MIN_HEIGHT = "148px";
 
@@ -27,6 +28,7 @@ interface Props {
   inventory: Inventory;
   inventoryItems: InventoryItemName[];
   onClick: (item: InventoryItemName) => void;
+  isFarming?: boolean;
 }
 
 const TAB_CONTENT_HEIGHT = 384;
@@ -40,36 +42,13 @@ export const InventoryTabContent = ({
   inventory,
   inventoryItems,
   onClick,
+  isFarming,
 }: Props) => {
   const { ref: itemContainerRef, showScrollbar } =
     useShowScrollbar(TAB_CONTENT_HEIGHT);
   const [scrollIntoView] = useScrollIntoView();
-  const categories = Object.keys(tabItems) as InventoryItemName[];
+  const categories = getKeys(tabItems) as InventoryItemName[];
   const [isTimeBoosted, setIsTimeBoosted] = useState(false);
-
-  useEffect(() => {
-    const firstCategoryWithItem = categories.find(
-      (category) => !!inventoryMapping[category]?.length
-    );
-
-    const defaultSelectedItem =
-      firstCategoryWithItem && inventoryMapping[firstCategoryWithItem][0];
-
-    if (defaultSelectedItem) {
-      setDefaultSelectedItem(defaultSelectedItem);
-    }
-  }, []);
-
-  useEffect(
-    () =>
-      setIsTimeBoosted(
-        hasBoost({
-          item: selectedItem as InventoryItemName,
-          inventory,
-        })
-      ),
-    [inventory, selectedItem]
-  );
 
   const inventoryMapping = inventoryItems.reduce((acc, curr) => {
     const category = categories.find(
@@ -84,6 +63,34 @@ export const InventoryTabContent = ({
 
     return acc;
   }, {} as Record<string, InventoryItemName[]>);
+
+  useEffect(() => {
+    const firstCategoryWithItem = categories.find(
+      (category) => !!inventoryMapping[category]?.length
+    );
+
+    const currentSelected = isFarming && getShortcuts()[0];
+
+    const defaultSelectedItem =
+      currentSelected ||
+      // Fallback for when a no active item selected
+      (firstCategoryWithItem && inventoryMapping[firstCategoryWithItem][0]);
+
+    if (defaultSelectedItem) {
+      setDefaultSelectedItem(defaultSelectedItem);
+    }
+  }, []);
+
+  useEffect(
+    () =>
+      setIsTimeBoosted(
+        hasBoost({
+          item: selectedItem as InventoryItemName,
+          inventory,
+        })
+      ),
+    []
+  );
 
   const findIfItemsExistForCategory = (category: string) => {
     return getKeys(inventoryMapping).includes(category);
@@ -106,6 +113,8 @@ export const InventoryTabContent = ({
   const inventoryIsEmpty = Object.values(inventoryMapping).every(
     (value) => value.length === 0
   );
+
+  console.log({ selectedItem });
 
   return (
     <div className="flex flex-col">
