+++ src/features/visiting/hud/Hud.tsx
@@ -1,12 +1,12 @@
 import React, { useContext } from "react";
 
-import { Inventory } from "./components/Inventory";
 import { AudioPlayer } from "components/ui/AudioPlayer";
 import { Menu } from "./components/Menu";
 import { Context } from "features/game/VisitingProvider";
 import { useActor } from "@xstate/react";
 import { VisitBanner } from "components/ui/VisitBanner";
 import { Balance } from "./components/Balance";
+import { Inventory } from "components/ui/Inventory";
 
 /**
  * Heads up display - a concept used in games for the small overlayed display of information.
@@ -22,7 +22,7 @@ export const Hud: React.FC = () => {
     <div data-html2canvas-ignore="true" aria-label="Hud">
       <Menu />
       <Balance />
-      <Inventory />
+      <Inventory inventory={gameState.context.state.inventory} />
       <AudioPlayer isFarming />
       <VisitBanner id={visitingLandId} />
     </div>
